+++ source/duke3d/src/savegame.h
@@ -114,7 +114,7 @@ extern bool g_saveRequested;
 extern savebrief_t * g_quickload;
 
 extern menusave_t * g_menusaves;
-extern size_t g_nummenusaves;
+extern uint32_t g_nummenusaves;
 
 int32_t sv_updatestate(int32_t frominit);
 int32_t sv_readdiff(int32_t fil);
